создаёт и подключает модуль create-photo
создаёт и подключает модуль для создания фотографии
